refactor: changed the definition of QueryPlanCost to f64
- The cost function has an arbitrary precision and can overflow i64.

+++ apollo-federation/src/query_plan/mod.rs
@@ -18,7 +18,7 @@ mod optimize;
 pub mod query_planner;
 pub(crate) mod query_planning_traversal;
 
-pub type QueryPlanCost = i64;
+pub type QueryPlanCost = f64;
 
 #[derive(Debug, Default)]
 pub struct QueryPlan {
